Restructure the packages to remove tangles.
- Split of processing and io from the topmost package
- export new packages
- rename OutputCellService to OutputDataRowService to be consistent with InputDataRowService.
