+++ org.knime.scijava.commands/META-INF/MANIFEST.MF
@@ -26,6 +26,7 @@ Export-Package: org.knime.knip.scijava.commands.impl,
  org.knime.scijava.commands,
  org.knime.scijava.commands.adapter,
  org.knime.scijava.commands.adapter.basic,
+ org.knime.scijava.commands.io,
  org.knime.scijava.commands.mapping,
  org.knime.scijava.commands.settings,
  org.knime.scijava.commands.settings.types,
