+++ org.knime.scijava.commands/src/org/knime/scijava/commands/mapping/ColumnToInputMappingService.java
@@ -1,5 +1,6 @@
 package org.knime.scijava.commands.mapping;
 
+import org.knime.scijava.commands.mapping.process.ColumnInputMappingKnimePreprocessor;
 import org.scijava.module.Module;
 
 /**
