+++ org.knime.scijava.commands/src/org/knime/scijava/commands/mapping/process/ColumnInputMappingKnimePreprocessor.java
@@ -1,13 +1,14 @@
-package org.knime.scijava.commands.mapping;
+package org.knime.scijava.commands.mapping.process;
 
 import org.knime.core.data.DataCell;
 import org.knime.core.data.DataRow;
 import org.knime.core.data.DataTable;
 import org.knime.core.data.DataTableSpec;
-import org.knime.scijava.commands.InputDataRowService;
-import org.knime.scijava.commands.KnimePreprocessor;
 import org.knime.scijava.commands.adapter.InputAdapter;
 import org.knime.scijava.commands.adapter.InputAdapterService;
+import org.knime.scijava.commands.io.InputDataRowService;
+import org.knime.scijava.commands.mapping.ColumnModuleItemMapping;
+import org.knime.scijava.commands.mapping.ColumnToInputMappingService;
 import org.scijava.Priority;
 import org.scijava.log.LogService;
 import org.scijava.module.Module;
